--- x-pack/plugins/cases/public/client/index.ts
@@ -1,15 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-export * from './api';
-export * from './helpers/can_use_cases';
-export * from './helpers/get_rule_id_from_event';
-export * from './ui/get_all_cases_selector_modal';
-export * from './ui/get_cases';
-export * from './ui/get_cases_context';
-export * from './ui/get_create_case_flyout';
-export * from './ui/get_recent_cases';
