+++ x-pack/plugins/cases/public/components/cases_context/cases_global_components.test.tsx
@@ -6,11 +6,9 @@
  */
 
 import React from 'react';
+import { getAllCasesSelectorModalNoProviderLazy } from '../../client/ui/get_all_cases_selector_modal';
+import { getCreateCaseFlyoutLazyNoProvider } from '../../client/ui/get_create_case_flyout';
 import { AppMockRenderer, createAppMockRenderer } from '../../common/mock';
-import {
-  getAllCasesSelectorModalNoProviderLazy,
-  getCreateCaseFlyoutLazyNoProvider,
-} from '../../client';
 import { getInitialCasesContextState } from './cases_context_reducer';
 import { CasesGlobalComponents } from './cases_global_components';
