+++ x-pack/plugins/cases/public/components/cases_context/cases_global_components.tsx
@@ -6,10 +6,8 @@
  */
 
 import React from 'react';
-import {
-  getAllCasesSelectorModalNoProviderLazy,
-  getCreateCaseFlyoutLazyNoProvider,
-} from '../../client';
+import { getAllCasesSelectorModalNoProviderLazy } from '../../client/ui/get_all_cases_selector_modal';
+import { getCreateCaseFlyoutLazyNoProvider } from '../../client/ui/get_create_case_flyout';
 import { CasesContextState } from './cases_context_reducer';
 
 export const CasesGlobalComponents = React.memo(({ state }: { state: CasesContextState }) => {
