+++ x-pack/plugins/cases/public/types.ts
@@ -24,16 +24,13 @@ import { CasesByAlertId, CommentRequestAlertType, CommentRequestUserType } from 
 import { UseCasesAddToExistingCaseModal } from './components/all_cases/selector_modal/use_cases_add_to_existing_case_modal';
 import { CreateCaseFlyoutProps } from './components/create/flyout';
 import { UseCasesAddToNewCaseFlyout } from './components/create/flyout/use_cases_add_to_new_case_flyout';
-
-import type {
-  CasesOwners,
-  GetAllCasesSelectorModalProps,
-  GetCasesProps,
-  GetCreateCaseFlyoutProps,
-  GetRecentCasesProps,
-  GetCasesContextProps,
-  getRuleIdFromEvent,
-} from './client';
+import type { CasesOwners } from './client/helpers/can_use_cases';
+import { getRuleIdFromEvent } from './client/helpers/get_rule_id_from_event';
+import type { GetCasesContextProps } from './client/ui/get_cases_context';
+import type { GetCasesProps } from './client/ui/get_cases';
+import { GetAllCasesSelectorModalProps } from './client/ui/get_all_cases_selector_modal';
+import { GetCreateCaseFlyoutProps } from './client/ui/get_create_case_flyout';
+import { GetRecentCasesProps } from './client/ui/get_recent_cases';
 
 export interface CasesPluginSetup {
   security: SecurityPluginSetup;
